pyproject.toml
[tool.poetry]
name = "booking"
version = "0.1.0"
description = ""
authors = ["Maksym Sydorchuk <maximsidorchuk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.4"
fastapi = "0.112.0"
uvicorn = "0.30.5"
sqlalchemy = "2.0.31"
alembic = "1.13.2"
asyncpg = "0.29.0"
python-jose = "3.3.0"
passlib = "1.7.4"
email-validator = "2.2.0"
pydantic-settings = "2.4.0"
fastapi-cache2 = {extras = ["redis"], version = "0.2.2"}
python-json-logger = "2.0.7"
greenlet = "3.0.3"
jinja2 = "3.1.4"
python-multipart = "0.0.9"
celery = "5.4.0"
flower = "2.0.1"
pillow = "10.4.0"
sqladmin = "0.19.0"
itsdangerous = "2.2.0"
pytest = "8.3.3"
httpx = "0.27.2"
pytest-asyncio = "0.24.0"
black = "24.8.0"
flake8 = "7.1.1"
autoflake = "2.3.1"
isort = "5.13.2"
pyright = "1.1.381"
sentry-sdk = {extras = ["fastapi"], version = "2.14.0"}
prometheus-fastapi-instrumentator = "7.0.0"
gunicorn = "23.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 88
target-version = ['py312']

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
line_length = 88

[tool.autoflake]
check_diff = true
imports = ["sqlalchemy", "app"]

[tool.pyright]
include = ["app"]

